+++ Console/Item.cs
@@ -5,9 +5,9 @@ namespace vendingMachine
    public class Item
     {
         public string name;
-        public double price;
+        public decimal price;
 
-        public Item(string itemName, double itemPrice)
+        public Item(string itemName, decimal itemPrice)
         {
             name = itemName;
             price = itemPrice;
